Rename getDateInterval
