+++ packages/lib-user/src/utils/index.js
@@ -2,5 +2,5 @@ export { default as createPanoptesUserGroup } from './createPanoptesUserGroup.js
 export { default as dateRanges } from './dateRanges.js'
 export { default as deletePanoptesUserGroup } from './deletePanoptesUserGroup.js'
 export { default as getBearerToken } from './getBearerToken.js'
-export { default as getStatsQueryFromDateRange } from './getStatsQueryFromDateRange.js'
+export { default as getDateInterval } from './getDateInterval.js'
 export { default as updatePanoptesUserGroup } from './updatePanoptesUserGroup.js'
